Fix bug where video_unit would only use the last bit for RGB
Fix a bug where video_unit only connected the last bit of the RGB output of objects_mux
to the input into VGA_Controller, causing the pixles to only be blue or black.

Also in this commit, to help debug this, changed all the arrays from unpacked arrays to
packed arrays (from "logic a [0:1]" to "logic [0:1] a").