fix all fixable warnings
I went over all the warnings, and found out how to make each of them disappear.
Now there should only be 5 messages that are not possible to remove - they are about:
- pins that are stuck at VCC/GND - this is a good thing, and there is no way to suppress the warning.
- connectivity issues - there are some instances that not all of their inputs and outputs are used, which is not a problem in most cases.
- some warning about the locked argument in the clock divider - this is a verilog file that I don't want to touch.

When only running analysis and synthesis, only 2 warnings about the connectivity and the stuck pins should appear.

From now on, any new code should not add warnings to the compilation, and we should look at the warnings for every compilation to catch errors.