fix bugs in interaction between player movement and borders
Fix the bug where the player sometimes could pass through the border.

Fix the bug where while touching the bottom border, the player could not
move left. This fix was done by changing the hit_colors variable that
controls what pixel of the spaceship counts as what side.
This was not a systematic fix, and it is a fix in an auto-generated file.
If we ever create a new spaceShipBitMap, we will need to fix this again.