fix player interaction with bottom border in spaceship bitmap
As statued in commit 4e4e21f9f4be9915a4740ed5371fb89e874b46d1, the fix
for the bad interaction between the player spaceship and the bottom border
is a bug in the spaceship bitmap file. After changing the spaceship bitmap
file at the last commit, the aforementioned bug returned, so this commit
redoes this fix.