abandoned code to implement missiles using RAM instead of FF
This was a test to use RAM instead of FF for the missiles. It worked
for a single missile, but we mistakingly thought that we have more
than one clock cycle per pixel.
The original idea to keep everything a missile needs in RAM was
scrubbed, but we might still use the RAM module implemented here for
something else, like keeping only the data that is not needed for
every pixel (like speed), and have only some things be in the RAM.